Implement new simplified character choice screen
The system of 4-directional movement and selection was too clunky for
something intended to be fun and fast. This is a much simplified system
where you only have to press a direction to choose an animal. It's now
also possible to undo; you're not locked in to an accidental choice.